Seperate save indicator without rerendering entire Editor Header
